Traspaso completo a JPA

+++ demo/src/main/java/com/runasagrada/demo/service/HotelServiceService.java
@@ -5,11 +5,11 @@ import java.util.Collection;
 import com.runasagrada.demo.entities.HotelService;
 
 public interface HotelServiceService {
-    public HotelService searchById(Integer id);
+    public HotelService searchById(Long id);
 
     public Collection<HotelService> getAllServices();
 
     public void save(HotelService service);
 
-    public void delete(Integer id);
+    public void delete(Long id);
 }
